+++ wserver/src/main/java/com/wanderdrop/wserver/model/Comment.java
@@ -15,7 +15,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private int commentId;
+    private Long commentId;
 
     @Column(name = "comment_heading", nullable = false)
     private String commentHeading;
